Add a "calculated panel" column to a dataset

Add a "calculated panel" column to a dataset

Usage

panel_lazy(plot_fn, data = dplyr::pick(dplyr::everything()))

Arguments

plot_fn

A function that produces a panel from a given subset of data, taking as arguments any variable names in data that you would like to be available in the plottinf function.

data

A data frame from which subsets will be extracted and plots will be made. Should be a superset of the summary dataset to which this plot column is being added.