Write panels

Write panels

Usage

write_panels(
  trdf,
  width = 500,
  height = 500,
  format = "png",
  force = FALSE,
  ...
)

Arguments

trdf

A trelliscope data frame created with as_trelliscope_df() or a data frame which will be cast as such.

width

Width in pixels of each panel.

height

Height in pixels of each panel.

format

The format of the image if it is not an htmlwidget. Can be either "png" or "svg".

force

Should the panels be forced to be written? If FALSE, the content of the panel column along with the width, height, and format parameters will be used to determine if the panel content matches panels that have already been written, in which case writing the panels will be skipped.

...

Other arguments passed to the plotting functions (either svglite::svglite or grDevices::png).

Note

The size of panels will vary when shown in the viewer, but here the specification of height and width help determine the plot aspect ratio as well as the initial resolution to render plot text, etc. with.

Examples

if (FALSE) {
library(ggplot2)

panel_dat <- (
  ggplot(gapminder, aes(year, lifeExp)) +
    geom_point() +
    facet_panels(~country + continent)
  ) |>
  nest_panels()

disp <- panel_dat |>
  as_trelliscope_df(name = "life_expectancy", path = tempfile()) |>
  write_panels(width = 800, height = 500, format = "svg")
}